Trait pc_keyboard::ScancodeSet

pub trait ScancodeSet {
    // Required method
    fn advance_state(&mut self, code: u8) -> Result<Option<KeyEvent>, Error>;
}

A mechanism to convert bytes from a Keyboard into KeyCode values.

This conversion is stateful.

Required Methods

fn advance_state(&mut self, code: u8) -> Result<Option<KeyEvent>, Error>

Handles the state logic for the decoding of scan codes into key events.